backend/requirements.txt
fastapi
uvicorn
pymongo
python-dotenv
pydantic[email]
passlib[argon2]
python-jose[cryptography]
elevenlabs
cloudinary
google-generativeai